functions/firebase/requirements.txt
functions-framework>=3.0.0
google-cloud-firestore>=2.14.0
google-cloud-aiplatform>=1.38.0
